+++ src/helpers/index.ts
@@ -4,3 +4,4 @@ export * from './slotContent';
 export * from './styles';
 export * from './sanitizer';
 export * from './blockWrapper';
+export * from './dom';